+++ lab1/src/Word.java
